feat(embed): first-class per-file update/remove through the admission seam
Task-board #24 / AAP ask 3. The single-file admission + publication seam
(ReindexResult, maybe_reindex, read_and_index*, admit_and_index_single_
path) relocates VERBATIM from `watcher` (server-only, notify-coupled) to
the new `live_index::single_file` (compiled under server AND embed); the
watcher keeps its exact call surface through a re-export, so every
watcher/freshen/reconciliation path is behavior-identical.

On top of the relocated seam, two embedder entry points, exported on
`symforge::embed` and pinned in the semver contract test:

- update_file_from_disk(shared, repo_root, relative_path): re-index one
  file through the SAME seam bulk load and the watcher use - scope/
  gitignore exclusions, metadata-first scout, secret-content demotion,
  content-hash-gated parse, generation-fenced publication. No retries,
  no sleeps (deadline-friendly for PID-1 embedders); ENOENT returns
  NotFound immediately.
- remove_file(shared, relative_path): generation-fenced removal; true =
  applied under the current generation (including the untracked no-op),
  false only when a concurrent retarget invalidated the fence.

This replaces the embedder-side hand-rolled process_file +
IndexedFile::from_parse_result + store-poking reindex, which bypassed
admission entirely - a secret-bearing edit could re-enter the index
through that lane; through the seam it is demoted exactly like the
watcher would.

Tests: facade round trip (new file admitted+parsed, unchanged file
hash-skipped, edit re-parsed, fenced removal, honest NotFound). Gates:
fmt, clippy -D warnings (full + embed), embed lib suite (1295), full
serial suite all targets green, release build.
